Refactor map hooks import paths and move hooks file
Updated import paths for map-related hooks to reference 'features/map/hooks' instead of nested Map directories. Renamed and moved 'hooks.js' to 'hooks/index.js' to simplify and centralize hook imports across the codebase.
