+++ src/features/map/components/Map/EditableMap.jsx
@@ -6,7 +6,7 @@ import { Button } from 'antd';
 import { GeoJsonLayer } from 'deck.gl';
 import { MapboxOverlay } from '@deck.gl/mapbox';
 
-import { useCameraForBounds, useFetchBuildings } from './hooks';
+import { useCameraForBounds, useFetchBuildings } from '../../hooks';
 import { calcBoundsAndCenter, EMPTY_FEATURE } from './utils';
 
 import { Map, useControl } from 'react-map-gl/maplibre';
